Fix build errors and revert single-implementation interfaces
- Remove single-implementation interfaces (IShadowMapProvider,
  IHeightProvider, IRenderTargetProvider, IFrameBuffers) as they
  don't provide value with only one implementation each
- Fix TerrainConfig.h include in WaterSystemGroup.h (forward declare)
- Fix ambiguous recordShadowDraw call in ShadowPassRecorder.cpp
- Fix WaterSystemGroup.cpp to use waterGroup().hasGBuffer() pattern
- Add missing RenderingInfrastructure.cpp and DescriptorInfrastructure.cpp
  to CMakeLists.txt

+++ src/core/passes/ShadowPassRecorder.cpp
@@ -46,7 +46,7 @@ void ShadowPassRecorder::record(VkCommandBuffer cmd, uint32_t frameIndex, float 
         (void)lightMatrix;  // Grass uses cascade index only
         if (config_.perfToggles && config_.perfToggles->grassShadows) {
             resources_.profiler->beginGpuZone(cb, "Shadow:Grass");
-            resources_.vegetation.grass().recordShadowDraw(cb, frameIndex, time, cascade);
+            resources_.vegetation.grass().recordShadowDraw(cb, frameIndex, time, static_cast<int>(cascade));
             resources_.profiler->endGpuZone(cb, "Shadow:Grass");
         }
     };
